OCC-4401 Use CmsService to get PersonalizationScriptComponent and build personalization context based on that data. Changes after review.
