+++ projects/core/src/personalization/http-interceptors/index.ts
@@ -3,7 +3,6 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { OccPersonalizationIdInterceptor } from './occ-personalization-id.interceptor';
 import { OccPersonalizationTimeInterceptor } from './occ-personalization-time.interceptor';
-import { OccPersonalizationContextInterceptor } from './occ-personalization-context.interceptor';
 
 export const interceptors: Provider[] = [
   {
@@ -14,11 +13,6 @@ export const interceptors: Provider[] = [
   {
     provide: HTTP_INTERCEPTORS,
     useExisting: OccPersonalizationTimeInterceptor,
-    multi: true,
-  },
-  {
-    provide: HTTP_INTERCEPTORS,
-    useExisting: OccPersonalizationContextInterceptor,
     multi: true,
   },
 ];
