+++ projects/core/src/personalization/index.ts
@@ -1,2 +1,3 @@
 export * from './personalization.module';
 export * from './config/personalization-config';
+export * from './services/personalization-context.service';
